Tighten scope of Future creation.
In rare cases, the reply arrives before we've set up our local data
structures so it gets put in a temporary location. Before this commit,
we always built a Future and the function that manipulates it -- even
if we've already received the reply.

In this commit, the infrastructure is set up only if the reply hasn't
yet arrived.