fix: properly handle BASIC STATUS and alarm blocks
When an `ItemData` is created for incoming data, it unconditionally read
the `.data` field of the reply. `BasicStatus_reply`,
`AnalogAlarm_reply`, and `DigitalAlarm_reply` objects don't have that
field so an exception is raised when processing them.

This commit sets the `.data` field of the `ItemData` structure to a
dictionary containing the data from these three reply types.