feat: Add SCALE service.
This commit adds the `acsys.scaling` module to the `acsys` package. The
SCALE service takes a DRF request and a raw, binary buffer and return an
array of scaled values (using the scaling information in the database
associated with the device.) The SCALE service also provides an
"unscaling" request to convert floating point into a raw buffer.

This service is mainly intended for library authors to support devices
that return structured data, which is impossible to scale/unscale
through the normal DPM channels.