change: trust caller's `node` parameter
In the previous version, if the caller specified a node to be used for
the SCALE service, it would still do a service discovery to that node to
make sure it was available. This meant scripts that did extra work to
remember their SCALE service were still paying the price of two ACNET
requests per conversion.

In this version, if a node is specified, the library trusts it and
simply sends the conversion request.