fix: :bug: correctly end gen in `.request_stream()`.
When exiting the generator that was returned by `.request_stream()`,
trying to cancel the ACNET request resulted in an "ignored exception"
warning. Canceling a request was an asynchronous function so, while we
were waiting for it to complete, the syste, continued to close out the
generator.

This commit waits for the cancel to complete before returning control.