TCP transport (#32)
* refactor: move documentation.

The property fields had docstrings associated with them. However, the
recommended way to document them is to add it to the class docstring.
This approach also works when properties are generated using the `attr`
package.

* fix: mention some APIs no longer use ACNET.

* refactor: remove `available_dpms()` and `find_dpm()`.

In the TCP world, these functions aren't necessary (and can't be
implemented because the service discovery is done by another layer.)

* refactor: the DPM module now uses the TCP interface.

* fix: use module name for logger instance.

* fix: move the port back to 6802.

Charlie had moved it to 6805 for testing but the final port address is
6802.

* refactor: optimize `consumeRawInt()`.

Profiling showed this function was called the most. This commit employs
an optimization which made it 18% faster. Instead of looping with the
iterator, we use `islice()` to quickly transfer the data to a buffer and
then use `int.from_bytes()` to convert it.

* refactor: use `next()` builtin.

Instead of calling `.__next__()`, which looks ugly, use the proper call
of `next()` (which then calls `.__next__()`.)

* refactor: no need to range check.

No need to range-check the result since `consumeRawInt` only accepts
integers up to 8 bytes in size.

* fix: add a `port` parameter to `DpmContext`.

The operational proxy is at port 6802 on *acsys-proxy*. This is now the
default configuration. When we want to test a new proxy, it'll run on
another port of the same machine. This parameter allows us to easily
change a script to use the development proxy.

* optimize: Optimize and use `unmarshal_int16()`.

`unmarshal_int16()` used to call the more-general `consumeRawInt()` and
then would check the range. It now reads the tag field and decodes one
of the two possible ways the decoding could go.

All routines which unmarshal 16-bits integers (field name decoding and
message name decoding) now call this routine.