fix: :children_crossing: Remove the request_reply timeout.
The request_reply timeout was requiring users to account for the case
when a response took too long. We've removed the timeout and thus
adopted the default ACNET timeout of 6(?) minutes. This should make
the interface for users simpler because they don't have to worry about
handling the timeout case. If they need to timeout quickly, they can
manually close the request.

This is an edge case and shouldn't be something encountered often, but
when it happens, it's annoying.

Also, add MIT LICENSE.

Bump to v0.12.2.