fix: :bug: Use custom request_reply for auth step.
The generalized _request function assumes that the response has a
status field. In the auth_step function the request responds with
a message that does not have a status field. So, we call request_reply
by hand. This means we need to remember to change this code when
changes are made to any other request_reply functions.

Bump to v0.12.3.