fix: :bug: Convert `dict_items` to `list` to allow for iteration.
With the previous changes, when the connection closes we automatically
retry connecting. In this code path, the stored requests state is
restored on reconnect. The dictionary items were passed to
`add_entries`, but `dict_items` has it's own type that's not
subscriptable. I converted the `dict_items` to a `list` to conform to
how `add_entries` works.