build: :construction_worker: Copy-paste build job to validate functionality.
I want to validate that I understand the problem with build by copy and
pasting the existing build and modifying Python 3.6.8 to match with a
valid version linux. While keeping the existing build without 3.6.8.

If this works, I'd like for a future push to clean this up with
conditional blocks or something of the sort.